Serial communication exception handling/simulation mode
Actuators are put into simulation mode if unable to connect to serial ports. Otherwise, lengths are initialized to -1 which stops the targets from being sent to the controllers. Upon receiving a length update from the controller, target is set to starting length therefore the house will not move on startup.